+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - New struct `direct::Timeline` that is like `tweet::Timeline`, but for DMs
 - New function to load a single direct message (`show`)
 - New function to load DMs received by the authenticated user (`received`)
+- New function to load DMs sent by the authenticated user (`sent`)
 
 ## [0.5.0] - 2016-10-02
 ### Added